+++ src/main/java/com/readme/novels/episodes/controller/EpisodesController.java
@@ -4,6 +4,7 @@ import com.readme.novels.episodes.dto.EpisodesDtoByUser;
 import com.readme.novels.episodes.dto.PlusViewsKafkaDto;
 import com.readme.novels.episodes.messagequeue.EpisodesKafkaProducer;
 import com.readme.novels.episodes.responseObject.ResponseEpisodesUser;
+import com.readme.novels.episodes.service.EpisodeHistoryService;
 import com.readme.novels.episodes.service.EpisodesService;
 import com.readme.novels.commonResponseObject.CommonDataResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -24,6 +25,7 @@ public class EpisodesController {
 
     private final EpisodesService episodesService;
     private final EpisodesKafkaProducer episodesKafkaProducer;
+    private final EpisodeHistoryService episodeHistoryService;
 
     @Operation(summary = "에피소드 조회", description = "에피소드 조회, 조회할 에피소드 id url 전달", tags = {"에피소드"})
     @ApiResponses({
@@ -44,7 +46,7 @@ public class EpisodesController {
         episodesKafkaProducer.plusViewCount("plusViewCount", plusViewsKafkaDto);
 
         // 최근 읽은 목록에 추가
-        if (!uuid.equals("")) { episodesService.addEpisodeHistory(id, uuid); }
+        if (!uuid.equals("")) { episodeHistoryService.addEpisodeHistory(id, uuid); }
 
         return ResponseEntity.ok(
             new CommonDataResponse<>(
