+++ src/main/java/com/readme/novels/episodes/service/EpisodeHistoryService.java
@@ -1,0 +1,6 @@
+package com.readme.novels.episodes.service;
+
+public interface EpisodeHistoryService {
+
+    void addEpisodeHistory(Long id, String uuid);
+}
