+++ src/main/java/com/readme/novels/episodes/service/EpisodesService.java
@@ -20,5 +20,4 @@ public interface EpisodesService {
 
     void plusViewsCount(Long id, Integer plusCount);
 
-    void addEpisodeHistory(Long id, String uuid);
 }
